ID-11.2: Add the ID name of the chromosomes to the whole_chromosomes table and avoid using the hardcoded chromosome ID when loading a RM repeat
Solution implemented and cleaning some code, as arguments passing to functions, sequence name, refseq accession number and organism passing as arguments in functions in command.py and analyze.py files.

The goal was to try to extract that data from the Sequence object.

Refseq accession number is not being sent to region_chromosomes table.