ID-13.4: Graph the results of ID-13.3 for a whole organism
I have to graph the results of recursively found repeats that are stored in the database for a whole organism and create the command to execute it.

This is already implemented for a single sequence. The SQL query must be changed to search by organism instead of searching by refseq accession number.
The "frequency" is taken from "largest_value" column from recursive_repeats_whole_chromosomes.
This is not real frequency, but the largest values (peaks) in the CGR 3D graph.