ID-28: [Refactor][MFA] Refactor the MFA analysis so that the mi_grids be extracted from DB (for whole analysis)
Instead of using the mi_grids found that are in RAM memory. Extract the mi_grids from the DB and perform the MFA analysis as a separate operation from CGR execution and extracting mi_grid from database.
Create additional methods, don't delete the current ones.
Use the mi_grid saved in database if it already was created, execute the CGR algorithm and insert it into the database if it has not been inserted before.

Finding recursive kmers and executing analysis for regions is missing.

Created a RegionMiGridsService, separated MiGrids table into whole and regions, changed flow in services, added update method for DBConnectionManager.