ID-29: [Automation] Automate the execution of RepeatMasker by Docker and WSL
WSL and Windows were not integrated since for each genome execution, the dfam -lib or -species must be specified so it has to be done manually. Two bash scripts were created to be executed from the Docker container where the results for the genome chromosomes are generated using an INPUT DIR (were fna sequences are located) and OUTPUT DIR (where RM results are saved) and a third parameter (species or lib accordingly).