ID-57: [Graph][Execute] Execute the commands for linear repeats algorithm only found in genes
Execute the command for finding and loading the results of the linear repeats algorithm into the database.
Execute the command for graphing the linear repeats algorithm from the results in the database.

Comment: Implemented finding kmers for repeats and gene repeats, graphing from file and from database and using sequence and genome execution.